+++ src/backend/equinoxe_hlc.ml
@@ -22,38 +22,8 @@
 (*                                                                           *)
 (*****************************************************************************)
 
-module Client = Http_lwt_client
+module Backend = Terminus_hlc
+(* This module is just a linking to the new package. It
+   is depracated and will be removed at some point. *)
 
-module Backend = struct
-  type nonrec 'a io = ('a, [ `Msg of string ]) Lwt_result.t
-
-  let return = Lwt_result.return
-  let map = Lwt_result.map
-  let bind f m = Lwt_result.bind m f
-  let fail e = Lwt_result.fail e
-
-  (***** Helpers *****)
-
-  let get_body = function
-    | Ok (_, None) -> return ""
-    | Ok (_, Some body) -> return body
-    | Error e -> fail e
-
-  let ( =<< ) f m = Lwt.bind m f
-
-  (**** Http methods ****)
-
-  let get ~headers ~url =
-    get_body =<< Client.one_request ~meth:`GET ~headers url
-
-  let post ~headers ~url body =
-    get_body =<< Client.one_request ~meth:`POST ~headers ~body url
-
-  let put ~headers ~url body =
-    get_body =<< Client.one_request ~meth:`PUT ~headers ~body url
-
-  let delete ~headers ~url =
-    get_body =<< Client.one_request ~meth:`DELETE ~headers url
-end
-
-include Equinoxe.Make (Backend)
+include Equinoxe.Make (Terminus_hlc)